pyproject.toml
[project]
name = "ams-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django-cors-headers>=4.7.0",
    "djangorestframework>=3.15.2",
    "djangorestframework-simplejwt>=5.4.0",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
]
